added separate controller for permissions, added EditorListener
